feat 003: 장바구니 추가 기능, productId로 통일

+++ src/main/java/com/mysite/cuffee/cart/repository/CartItemRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface CartItemRepository extends JpaRepository<CartItem, Long> {
-    Optional<CartItem> findByCartIdAndProductId(int cartId, int productId);
+    Optional<CartItem> findByCartIdAndProductId(long cartId, long productId);
 }
